moved hash class to seperate file hash.py
